Rename protoyaml to parser
